% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_pearson.R
\name{cor.est}
\alias{cor.est}
\title{Correlation analysis in a RDS study}
\usage{
cor.est(x, y, data)
}
\arguments{
\item{x}{Numeric vector; a continuous variable x}

\item{y}{Numeric vector; a continuous variable y}

\item{data}{Data frame; Contains data from bootstrap_RDS function with boot_n and variables of interest}
}
\value{
A numeric scalar value indicating correlation coefficient between x and y
}
\description{
Correlation analysis in a RDS study
}
\examples{
cor_coef = cor.est(x, y, data)
}
