% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descr.stats.est}
\alias{descr.stats.est}
\title{Descriptive statistics in an RDS study}
\usage{
descr.stats.est(data, x)
}
\arguments{
\item{data}{Data frame; The output from bootstrap_RDS with boot_n and respondent ID}

\item{x}{Numeric vector; A variable of interest}
}
\value{
Data frame; Containes mean, median, variance and associated 95\% CIs
}
\description{
Descriptive statistics in an RDS study
}
\examples{
data('RDStoydata')

# Preprocess data with RDSdata function
rds_data <- RDSdata(data = RDStoydata,unique_id = "ID",
redeemed_coupon = "CouponR",
issued_coupon = c("Coupon1",
                 "Coupon2",
                 "Coupon3"),
               degree = "Degree",
               result = c('Age','Sex'))


# Run bootstrap_RDS with rds_data
results = bootstrap_RDS(RESPONDENT_ID = rds_data$ID, SEED_ID = rds_data$S_ID,
SEED = rds_data$SEED, RECRUITER_ID = rds_data$R_CP,
data.cov = rds_data[,c('Age', 'Sex')], type = 'boot_chain_one', n.times = 100,
return_data = T)

# Calculate descriptive statistics using results from bootstrap_RDS
out = descr.stats.est(results, as.numeric(results$Age))
out
}
