% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS_data_handling_weight.R
\name{RDSdata}
\alias{RDSdata}
\title{Data processing in an RDS study}
\usage{
RDSdata(
  data,
  unique_id,
  redeemed_coupon,
  issued_coupons,
  degree,
  zero_degree_method = "hotdeck",
  NA_degree_method = "hotdeck",
  result
)
}
\arguments{
\item{data}{Data frame; Should contain ID numbers for the nodes in the social network, corresponding redeemed coupon number, and issued coupon number}

\item{unique_id}{Character vector; The column name of the column in the data that represents the ID numbers for the nodes in the social network}

\item{redeemed_coupon}{Character vector; The column name of the column in the data that represents coupon numbers of coupons redeemed by respondents when participating in the survey}

\item{issued_coupons}{Character vector; The column name of the column in the data that represents the coupon numbers of coupons issued to respondents}

\item{degree}{Numeric vector;  The column name of the column in the data that represents the degree of the respondents}

\item{zero_degree_method}{Character; This parameter is used to set the method for imputing zero values in the 'degree' variable. Three methods are available for selection: mean, median, and hotdeck. If this parameter is not set, the default imputation method is hotdeck}

\item{NA_degree_method}{Character; This parameter is used to set the method for imputing missing values in the 'degree' variable. There are three methods to choose from: mean, median, and hotdeck. If this parameter is not set, the default method is hotdeck.}

\item{result}{Vector; Column names of columns representing the interview outcome data about the respondents that should be included in the output data}
}
\value{
\item{ID}{Character vector; Renamed unique_id variable}
\item{R_CP}{Character vector; Renamed redeemed coupon variable}
\item{T_CP1 - T_CPn}{Character vector(s); Renamed issued coupon variable}
\item{DEGREE}{Numeric vector; Renamed degree variable}
\item{WAVE}{Numeric vector; Indicates which round the node was introduced into the survey, and the value of Seed is 0}
\item{S_ID}{chacarter vector; Indicates the ID of the seed corresponding to the node. The value of the seed is itself}
\item{R_ID}{Character vector; Indicates the ID of the node who recruits the node joinning the survey. The value of seed is NA}
\item{SEED}{Numeric vector: Values are only 0 and 1, they are used to indicate whether the node is seed or not. If it is seed, the value is 1, if not, it is 0.}
\item{CT_T_CT}{Numeric data, indicates how many coupons have been issued to this node to invite others to join the survey}
\item{CT_T_CP_USED}{Numeric data, indicates how many coupons issued to this node have been used to invite others to join the survey.}
\item{DEGREE_MODIFY}{Numeric data, indicates the `degree` data after imputation of zero values and missing values.}
}
\description{
Data processing in an RDS study
}
\examples{
# for preprocessing use RDStoydata

data('RDStoydata')

rds_data <- RDSdata(data = RDStoydata,unique_id = "ID",
redeemed_coupon = "CouponR",
issued_coupon = c("Coupon1",
                 "Coupon2",
                 "Coupon3"),
               degree = "Degree",
               result = c('Age','Sex'))
}
