% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{reg.est.lm}
\alias{reg.est.lm}
\title{Linear regression modeling in an RDS study}
\usage{
reg.est.lm(data, formula, include.degree = F)
}
\arguments{
\item{data}{Data frame; Contains data from bootstrap_RDS function with boot_n and variables of interest}

\item{formula}{An object of class 'formula'; description of the model with dependent and independent variables. Note that for linear regression the dependent variable must be continuous. Use reg.est.glm to perform regression with binary categorical dependent variable.}

\item{include.degree}{Logical; when TRUE the function performs weighted linear regression with DEGREE used as weights}
}
\value{
A data frame consisting of the point estimates and associated standard errors
}
\description{
Linear regression modeling in an RDS study
}
\examples{
res_lm_out = reg.est.lm(data, formula, include.degree=G)

}
