% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descr.stats.est}
\alias{descr.stats.est}
\title{Descriptive statistics in an RDS study}
\usage{
descr.stats.est(data, x)
}
\arguments{
\item{data}{Data frame; The output from bootstrap_RDS with boot_n and respondent ID}

\item{x}{Numeric vector; A variable of interest}
}
\value{
Data frame; Containes mean, median, variance and associated 95\% CIs
}
\description{
Descriptive statistics in an RDS study
}
\examples{
out = descr.stats.est(data, x)

}
