\name{descriptive_stats}
\alias{descriptive_stats}

\title{
Descriptive statistics from the 
}

\description{
This function calculates descriptive statistics based on the provided results from bootstrap_RDS. FORMULA
}

\usage{
out = descr.stats.est(data, x)
}

\arguments{
  \item{data}{Data frame; The output from bootstrap_RDS with boot_n and respondent ID}
  \item{x}{Numeric vector; A variable of interest}
}

\value{
  \item{data}{Data frame; Containes mean, median, variance and associated 95\% CIs}
}

\author{
THE AUTHORS WILL BE PLACED HERE
}


\examples{
EXAMPLES WILL BE PLACED HERE
}

